Complete overhaul of fault heapsort. Added important comments. Added all jobs PRIOR to the start of the simulation to give meaning to ARRIVE_MIN, ARRIVE_MAX. Deleted unecessary methods like generate_event(), which could be completed in main(). Fixed massive FIFO queue bug due to not passing all cpu and disk structs by reference. However, Disk1 and Disk2 do not have proper functionality, and connections to CPU haven't been made due to overhaul. Additionally, fixed incorrect random_generator values and added probability function.

+++ read_file.h
@@ -4,7 +4,7 @@
 #include <stdlib.h>
 #include <string.h>
 
-int *read_file();
+float *read_file();
 
 #endif //READ_FILE_H
